chore(*): add module export Правки:  - добавил экспорт по модулям  - обновил библиотеки
