+++ package.json
@@ -2,8 +2,28 @@
   "name": "fork-ipc",
   "version": "1.0.1",
   "description": "Safe and easy fork IPC wrapper.",
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
+  "main": "./dist/index.js",
+  "exports": {
+    ".": "./dist/index.js",
+    "./parent": "./dist/parent.js",
+    "./child": "./dist/child.js",
+    "./system": "./dist/system.js",
+    "./package.json": "./package.json"
+  },
+  "types": "./dist/index.d.ts",
+  "typesVersions": {
+    "*": {
+      "parent": [
+        "./dist/parent.d.ts"
+      ],
+      "child": [
+        "./dist/child.d.ts"
+      ],
+      "system": [
+        "./dist/system.d.ts"
+      ]
+    }
+  },
   "scripts": {
     "build": "tsc && tsc-alias",
     "test": "jest",
@@ -29,9 +49,9 @@
   },
   "devDependencies": {
     "@types/jest": "^26.0.24",
-    "@types/node": "^16.3.2",
+    "@types/node": "^16.4.10",
     "jest": "^27.0.6",
-    "tsc-alias": "^1.3.7",
+    "tsc-alias": "^1.3.8",
     "typescript": "^4.3.5"
   }
 }
