+++ src/child.ts
@@ -1,0 +1,8 @@
+import { executeViaParent, servicesAnnouncement } from '@Lib/child';
+import { emitToParent } from '@Lib/emitter';
+
+export const child = {
+  emit: emitToParent,
+  execute: executeViaParent,
+  servicesAnnouncement: servicesAnnouncement
+};
