+++ src/index.ts
@@ -2,50 +2,15 @@
  * Fork IPC - safe and easy fork IPC wrapper.
  */
 
-import { executeViaParent, getChildrens, servicesAnnouncement } from '@Lib/child';
-import { emitToParent, onceFromChild, onFromChild, removeChildListener } from '@Lib/emitter';
-import { registerLocal } from '@Lib/localProcessor';
-import { allowToChild, getChildrensGrants, registerChild } from '@Lib/parent';
-import { getProcessors } from '@Lib/processor';
-import { doRequest, getRequestQueue } from '@Lib/requests';
-import { getIdConter } from '@Lib/utils';
+import { child } from '@/child';
+import { parent } from '@/parent';
+import { system } from '@/system';
 
-const child = {
-  emit: emitToParent,
-  execute: executeViaParent,
-  servicesAnnouncement: servicesAnnouncement
-};
+const { servicesAnnouncement } = child;
+const { registerChild, execute } = parent;
 
-const parent = {
-  allowToChild: allowToChild,
-  execute: doRequest,
-  on: onFromChild,
-  once: onceFromChild,
-  registerChild: registerChild,
-  registerLocal: registerLocal,
-  removeListener: removeChildListener
-}
+export { execute, registerChild, servicesAnnouncement }
+export { child, parent, system }
 
-/*
- * For diagnostic
- */
-const system = {
-  diagnostic: () => (
-    {
-      idCounter: getIdConter(),
-      childrens: getChildrens(),
-      childrensGrants: getChildrensGrants(),
-      processors: getProcessors(),
-      requestQueue: getRequestQueue()
-    }
-  )
-}
-
-export { doRequest as execute }
-export { registerChild }
-export { servicesAnnouncement }
-export { child }
-export { parent }
-export { system }
-
+/** @deprecated */
 export default { child, parent, system }
