+++ src/system.ts
@@ -1,0 +1,18 @@
+import { getChildrens } from '@Lib/child';
+import { getChildrensGrants } from '@Lib/parent';
+import { getProcessors } from '@Lib/processor';
+import { getRequestQueue } from '@Lib/requests';
+import { getIdConter } from '@Lib/utils';
+
+/*
+ * For diagnostic
+ */
+export const system = {
+  diagnostic: {
+    idCounter: getIdConter(),
+    childrens: getChildrens(),
+    childrensGrants: getChildrensGrants(),
+    processors: getProcessors(),
+    requestQueue: getRequestQueue()
+  }
+}
